requirements.txt
PyNaCl
ffmpeg
discord.py
discord.py[voice]
python-dotenv
responses